JS Contest R2
1. Media Editor

The editor is created using canvas.
- All modes are implemented: enhance, crop, draw, text, stickers.
- The history module is implemented and connected to all of the modes.
- In crop and rotate mode, the aspect-ratio is preserved both in the editor itself and in layers.
- All 3 types of stickers are supported, including Lottie and Video. Animated stickers are always played in the editor.
- Fonts in text are implemented using system fonts (see comment with details in code).
- Saving to GIF is currently disabled, but the file is in the codebase. Turned out to be an inefficient solution with a high threshold of false positive cases, so now only first frame is saved as file.

2. Multiaccount Mode

Ability to authorize additional accounts and store multiple sessions.

- Authorization is made in credentialless session and then stored along with all other keys. It is also checked for premium limits.
- Subscription to user updates is done in a separate channel and browser notifications are controlled there as well.

3. Scroll issues

_Not implemented_.
As far as I was able to understand, we can see a drop in FPS during scrolling on some devices.
In my opinion, this is not due to the rendering/paintint, but more to the way we manage memory.
This can be noticed not only on smartphones. If we work for a long time on a desktop, we can get a similar effect.
It is necessary to work more thoroughly on processes that work with memory:
use efficient data-structures, use `destroy` more often, profile all classes that do `MOUNT_CLASS_TO`, which also potentially affects memory consumption.